requirements.txt
openai
faiss-cpu
pdfplumber
pyyaml